refactor: separate newSession
