+++ src/Components/NewSession/NewSession.module.css
@@ -1,0 +1,20 @@
+.new_session {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  width: 515px;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  margin-top: 32px;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+@media screen and (max-width: 750px) {
+  .new_session {
+    width: 440px;
+  }
+}
+/*# sourceMappingURL=NewSession.module.css.map */
