+++ src/Components/NewSession/NewSession.module.scss
@@ -1,0 +1,10 @@
+.new_session {
+    display: flex;
+    width: 515px;
+    justify-content: space-between;
+    margin-top: 32px;
+    align-items: center;
+    @media screen and (max-width: 750px) {
+        width: 440px;
+    }
+}
