+++ src/pages/MainPage/MainPage.module.scss
@@ -25,16 +25,7 @@
         padding: 20px 10px 0;
     }
 }
-.new_session {
-    display: flex;
-    width: 515px;
-    justify-content: space-between;
-    margin-top: 32px;
-    align-items: center;
-    @media screen and (max-width: 750px) {
-        width: 440px;
-    }
-}
+
 .poker_planning {
     display: block;
     margin: 0 auto;
